split test files and add mocks to some tests
