+++ testRequirements.txt
@@ -1,3 +1,5 @@
 LMIPy
 pytest-cov==2.7.1
 codecov==2.0.15
+requests_mock==1.7.0
+pytest-env==0.6.2
