+++ tests/image_tests.py
@@ -1,0 +1,23 @@
+import os.path
+
+from LMIPy import ImageCollection
+
+try:
+    API_TOKEN = os.environ.get("API_TOKEN", None)
+except:
+    raise ValueError(f"Failed to access keys for test.")
+
+
+# ----- ImageCollection Tests -----#
+
+def test_image_collection_search():
+    ic = ImageCollection(lon=28.271979, lat=-16.457814, start='2018-06-01', end='2018-06-20')
+    assert len(ic) > 0
+
+
+# ----- Image Tests -----#
+
+def test_create_image():
+    ic = ImageCollection(lon=28.271979, lat=-16.457814, start='2018-06-01', end='2018-06-20')
+    im = ic[0]
+    assert im.attributes['provider'] is not None
