+++ tests/test_assets/GET_7cf3fab2-3fbe-4980-b572-712207b2c8c7.json
@@ -1,0 +1,62 @@
+{
+  "data": {
+    "id": "7cf3fab2-3fbe-4980-b572-712207b2c8c7",
+    "type": "dataset",
+    "attributes": {
+      "name": "Template Dataset",
+      "slug": "Template-Dataset-CLONED_1",
+      "type": null,
+      "subtitle": null,
+      "application": [
+        "gfw"
+      ],
+      "dataPath": null,
+      "attributesPath": null,
+      "connectorType": "rest",
+      "provider": "cartodb",
+      "userId": "57a0aa1071e394dd32ffe137",
+      "connectorUrl": "https://wri-01.carto.com/tables/gfw_land_rights/public_map",
+      "sources": [],
+      "tableName": "gfw_land_rights",
+      "status": "saved",
+      "published": false,
+      "overwrite": false,
+      "verified": false,
+      "blockchain": {},
+      "mainDateField": null,
+      "env": "staging",
+      "geoInfo": false,
+      "protected": false,
+      "legend": {
+        "date": [],
+        "region": [],
+        "country": [],
+        "nested": [],
+        "integer": [],
+        "short": [],
+        "byte": [],
+        "double": [],
+        "float": [],
+        "half_float": [],
+        "scaled_float": [],
+        "boolean": [],
+        "binary": [],
+        "text": [],
+        "keyword": []
+      },
+      "clonedHost": {},
+      "errorMessage": "",
+      "taskId": null,
+      "createdAt": "2019-09-20T11:24:40.408Z",
+      "updatedAt": "2019-09-24T15:28:54.521Z",
+      "dataLastUpdated": null,
+      "user": {
+        "name": "Alicia",
+        "email": "alicia.arenzana@vizzuality.com",
+        "role": "ADMIN"
+      },
+      "widgetRelevantProps": [],
+      "layerRelevantProps": []
+    }
+  }
+}
