+++ tox.ini
@@ -12,3 +12,9 @@ passenv =
 commands =
     coverage run --source LMIPy setup.py test
     coverage report -m%
+
+[pytest]
+env =
+    API_TOKEN=testToken
+testpaths = tests
+python_files= **_tests.py
